+++ .gitignore
@@ -56,7 +56,5 @@ app/assets/builds/*
 /app/assets/builds/*
 !/app/assets/builds/.keep
 
-/gems/laa_multi_step_forms/tmp/*
-
 # Ignore puppeteer cache
 .cache
