+++ app/assets/stylesheets/application.scss
@@ -1,4 +1,4 @@
-@import "../../../gems/laa_multi_step_forms/app/assets/stylesheets/laa_multi_step_forms/base.scss";
+@import "./laa_multi_step_forms/base.scss";
 @import "./custom/govuk-date-legend";
 @import "./custom/govuk-formgroup";
 @import "./custom/govuk-summary-list";
