+++ app/assets/stylesheets/print.scss
@@ -1,1 +1,1 @@
-@import "../../../gems/laa_multi_step_forms/app/assets/stylesheets/laa_multi_step_forms/print.scss";
+@import "laa_multi_step_forms/print.scss";
