+++ app/services/decisions/base_decision_tree.rb
@@ -1,38 +1,74 @@
 module Decisions
   class BaseDecisionTree
     class InvalidStep < RuntimeError; end
+    WRAPPER_CLASS = SimpleDelegator
 
-    attr_reader :form_object, :step_name
+    class << self
+      def rules
+        @rules ||= {}
+      end
+
+      def from(source)
+        raise "Rule already exists for #{source}" if rules[source]
+
+        rules[source] = Rule.new
+      end
+    end
+
+    attr_reader :form_object, :step_name, :rule
 
     def initialize(form_object, as:)
       @form_object = form_object
       @step_name = as
-    end
-
-    def destination
-      raise 'implement this action, in subclasses'
+      @rule = self.class.rules[as]
     end
 
     delegate :application, :record, to: :form_object
 
+    def destination
+      return to_route(index: '/nsm/claims') unless rule
+
+      detected = nil
+      _, destination = rule.destinations.detect do |(condition, _)|
+        detected = condition.nil? || wrapped_form_object.instance_exec(&condition.to_proc)
+      end
+
+      return to_route(index: '/nsm/claims') unless destination
+
+      to_route(resolve_procs(destination, detected))
+    end
+
     private
 
-    # :nocov:
-    def index(step_controller, params = {})
-      { controller: step_controller, action: :index }.merge(params)
+    def wrapped_form_object
+      self.class::WRAPPER_CLASS.new(form_object)
     end
 
-    def show(step_controller, params = {})
-      url_options(step_controller, :show, params)
+    def resolve_procs(hash_or_proc, detected)
+      hash = resolve_proc(hash_or_proc, detected)
+      hash.transform_values { |value| resolve_proc(value, detected) }
     end
 
-    def edit(step_controller, params = {})
-      url_options(step_controller, :edit, params)
+    def resolve_proc(value, detected)
+      if value.respond_to?(:call)
+        value.arity.zero? ? wrapped_form_object.instance_exec(&value) : value.call(detected)
+      else
+        value
+      end
     end
 
-    def url_options(controller, action, params = {})
-      { controller: controller, action: action, id: application }.merge(params)
+    def to_route(hash)
+      if hash[:edit]
+        { controller: hash.delete(:edit), action: :edit }.merge(hash)
+      elsif hash[:show]
+        { controller: hash.delete(:show), action: :show }.merge(hash)
+      elsif hash[:index]
+        { controller: hash.delete(:index), action: :index }.merge(hash)
+      elsif hash[:new]
+        { controller: hash.delete(:new), action: :new }.merge(hash)
+      else
+        raise "No known verbs found in #{hash.inspect}"
+      end
     end
-    # :nocov:
   end
 end
