+++ spec/attributes/integer_time_period_spec.rb
