+++ spec/attributes/type/fully_validatable_decimal_spec.rb
