+++ spec/attributes/type/fully_validatable_integer_spec.rb
