+++ spec/attributes/type/gbp_spec.rb
