+++ spec/attributes/type/possibly_translated_array_spec.rb
