+++ spec/attributes/type/possibly_translated_string_spec.rb
