+++ spec/attributes/type/stripped_string_spec.rb
