+++ spec/attributes/type/time_period_spec.rb
