+++ spec/attributes/type/value_object_spec.rb
