+++ spec/lib/errors_spec.rb
