+++ spec/models/court_spec.rb
