+++ spec/presenters/task_list/collection_spec.rb
