+++ spec/presenters/task_list/section_spec.rb
