+++ spec/validators/time_period_validator_spec.rb
