+++ spec/value_objects/nsm/case_disposal_spec.rb
