+++ spec/value_objects/nsm/case_outcome_spec.rb
