+++ spec/value_objects/nsm/claim_type_spec.rb
