+++ spec/value_objects/value_object_spec.rb
