+++ spec/value_objects/yes_no_answer_spec.rb
